refactor(backend-service-user): move db modules
